requirements.txt
pillow==9.0.1
requests==2.23.0